Issue #11163: Enforce file size for CyclomaticComplexity input files
